+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## 1.1.0 (Unreleased)
+
+IMPROVEMENTS:
+
+- change the source of vpc module, and add `profile` parameter [GH-3]( https://github.com/terraform-alicloud-modules/terraform-alicloud-network-with-security-group/pull/3) 
+
 ## 1.0.0 (Feburary 20, 2020)
 
 - **NEW:**  `network-with-security-group` [GH-1]( https://github.com/terraform-alicloud-modules/terraform-alicloud-network-with-security-group/pull/1)